test: Set a strict package for generated Assertions class for AssertJ Assertions generator. -- See https://joel-costigliola.github.io/assertj/assertj-assertions-generator-maven-plugin.html#quickstart chore: Extracted more versions (not yet all) to common maven properties.
